Move `*.instruction.util` to `*.util`
