+++ core/src/main/java/me/gamercoder215/socketmc/util/render/Vertex.java
@@ -1,4 +1,4 @@
-package me.gamercoder215.socketmc.instruction.util.render;
+package me.gamercoder215.socketmc.util.render;
 
 import java.io.Serial;
 import java.io.Serializable;
